rename column_base_function to schema_base
