+++ rlink/src/functions/mod.rs
@@ -1,5 +1,5 @@
 pub mod broadcast_flat_map;
-pub mod column_base_function;
 pub mod keyed_state_flat_map;
 pub mod percentile;
 pub mod round_robin_flat_map;
+pub mod schema_base;
