+++ rlink/src/functions/schema_base/key_selector.rs
@@ -1,6 +1,6 @@
 use crate::api::element::Record;
 use crate::api::function::{Context, Function, KeySelectorFunction};
-use crate::functions::column_base_function::FunctionSchema;
+use crate::functions::schema_base::FunctionSchema;
 
 #[derive(Debug)]
 pub struct ColumnBaseKeySelector {
